Marshaller reset at startup & gets axis mac addrs from commander
THe marshaller is reset when the commander starts up to insure
everything is fresh and ready to gp. Commander also sends the
axis mac addresses to the marshaller at startup removing the need
to keep a local file synced with a copy of the file on the Rpi.
A framework was put in place to allow commands targeted at the
marshaller to be easily implemented. To add a command, create
the command on the commander side, or wherever it will come from,
add the name of the command to the cmd dictionary in the marshaller
and created a marshaller method of the same name that implements
the functionality needed to carry out the command, using the
function as a callbac value in the dictionary.